requirements.txt
# Core
fastapi
uvicorn[standard]
pydantic-settings
sqlalchemy
greenlet # Required for async sqlalchemy
asyncpg
alembic

# Auth & HTTP
httpx
python-jose[cryptography]
passlib[bcrypt]
cryptography

# Dev
pytest
pytest-asyncio
